+++ src/Resources/app/administration/src/init/api-service.init.js
@@ -4,6 +4,8 @@ import SwagPayPalApiCredentialsService
     from '../core/service/api/swag-paypal-api-credentials.service';
 import SwagPayPalPaymentService
     from '../core/service/api/swag-paypal-payment.service';
+import SwagPaypalPaymentMethodServiceService
+    from '../core/service/api/swag-paypal-payment-method.service';
 
 const { Application } = Shopware;
 
@@ -23,4 +25,10 @@ Application.addServiceProvider('SwagPayPalPaymentService', (container) => {
     const initContainer = Application.getContainer('init');
 
     return new SwagPayPalPaymentService(initContainer.httpClient, container.loginService);
+});
+
+Application.addServiceProvider('SwagPaypalPaymentMethodServiceService', (container) => {
+    const initContainer = Application.getContainer('init');
+
+    return new SwagPaypalPaymentMethodServiceService(initContainer.httpClient, container.loginService);
 });